% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb_to_xyz_array.R
\name{rgb_to_xyz_array}
\alias{rgb_to_xyz_array}
\title{customize the rgb_to_xyz function (from schemr) to work on arrays ##}
\usage{
rgb_to_xyz_array(rgb, transformation = "sRGB", linear_func = NULL)
}
\description{
customize the rgb_to_xyz function (from schemr) to work on arrays ##
}
