% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLogo.R
\name{SimpLogo}
\alias{SimpLogo}
\title{Convert protein sequence alignments to pseudo-1D colored representations.}
\usage{
SimpLogo(seq.dir, pattern = "*fa", group.names = NULL)
}
\arguments{
\item{seq.dir}{Directory containing one or more protein sequence alignments. If multiple alignments are present, they must be of equal lengths (i.e., a master alignment split into multiple groups).}

\item{pattern}{File extension for sequence alignment files (e.g., "*fa").}

\item{group.names}{Character vector containing names for sequence groups (1 per alignment file; make sure it's in the same order as your sequence files). If NULL, names are taken from filenames.}
}
\value{
Returns a formatted SimpLogo data table ready for plotting.
}
\description{
This function takes one or more sequence alignment file and converts it to a SimpLogo format for plotting.
}
