% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplogo_theme.R
\name{simplogo_main_theme}
\alias{simplogo_main_theme}
\title{define the simplogo_main_theme and simplogo_ic_theme functions for ggplot ##}
\usage{
simplogo_main_theme()
}
\description{
define the simplogo_main_theme and simplogo_ic_theme functions for ggplot ##
}
