% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLogoPlot.R
\name{SimpLogoPlot}
\alias{SimpLogoPlot}
\title{Plot the SimpLogo representation using ggplot2.}
\usage{
SimpLogoPlot(simplogo, plot.ic = TRUE, position.start = NULL)
}
\arguments{
\item{simplogo}{Formatted output of SimpLogo() function.}

\item{plot.ic}{Boolean whether to display position-wise information content.}

\item{position.start}{Number of first position residue (if using reference protein that doesn't start at position 1).}
}
\value{
Returns a ggplot2 object containing the SimpLogo representation.
}
\description{
This function is the plotting engine for SimpLogo. It takes the output of the SimpLogo() function and creates the graphical representation using ggplot2.
}
