% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLogo.R
\name{SimpLogo}
\alias{SimpLogo}
\title{Convert protein sequence alignments to pseudo-1D colored representations.}
\usage{
SimpLogo(seq.dir, pattern = "*fa", group.names = NULL)
}
\arguments{
\item{seq.dir}{Directory containing one or more protein sequence alignments. If multiple alignments are present, they must be of equal lengths (i.e., a master alignment split into multiple groups).}

\item{pattern}{File extension for sequence alignment files (e.g., "*fa").}

\item{group.names}{Character vector containing names for sequence groups (1 per alignment file; make sure it's in the same order as your sequence files). If NULL, names are taken from filenames.}

\item{res.colors}{Named vector of colors for each residue type. Must be named and of the same length as the total number of residue groupings (9 groups by default). The default color scheme is as follows: H/K/R = #4363d8, D/E = #e6194b, A/G = #2F4F4F, Q/N/S/T = #ffe119, M/V/L/I = #3cb44b, F/Y/W = #9a6324, P = purple, C = #46f0f0, - = NA.}
}
\value{
Returns a formatted SimpLogo data table ready for plotting.
}
\description{
This function takes one or more sequence alignment file and converts it to a SimpLogo format for plotting.
}
